moved custom hook in it's own file
