+++ client/src/hooks/useCallback.js
@@ -1,0 +1,12 @@
+import {useState} from 'react'
+
+//To think of better naming convention
+const useCallback = (input) => {
+    const [value, setValue] = useState(input)
+    function setCallBack(value) {
+        setValue(value)
+    }
+    return [value, setCallBack]
+}
+
+export {useCallback}
